requirements.txt
streamlit>=1.38.0
langchain>=0.3.25
langchain-community>=0.3.25
langchain-chroma>=0.2.4
langchain-groq>=0.2.0
sentence-transformers>=2.5.1
pymupdf>=1.26.0
chromadb>=0.5.11
numpy>=1.26.4
requests>=2.32.0
python-dotenv>=1.0.1
